Create a Play plugin to auto-reconfigure Play apps
To reconfigure a Play application, an application would use a Global
class which is driven by play.api.GlobalPlugin that is defined with
order 10000 in a play.plugins file inside the Play framework. To
achieve a similar effect to a Global class with a class name that will
not conflict with an application class, define a plugin class and give
it an order of 9999.

Remove auto-reconfiguration code from Bootstrap but Retain Bootstrap
while cloudfoundry-buildpack-java still uses it.

[#52912355]