Grails Auto-reconfiguration
Before the major changes in auto-reconfiguration Grails auto-
reconfiguration worked by virtue of it's use of a web.xml file.  After
the major changes it no longer worked because there were no known hook
points into the Grails lifecycle for adding the auto-reconfiguration
(without web.xml).  After a great meeting today with Graeme and Lari
it was suggested that we hook in with a Grails plugin.  After just a
couple of minutes work, this was in place and now all application
types that use Spring have auto-reconfiguration.

[#49695439]