Invalid Cloud Services
Previously, when the test to determine if cloud services were being
used directly was run, there was no initial check to see if the cloud
service classes could even be loaded.  In older versions of the
spring-cloud library there were instances where classes were listed as
cloud services, but those classes did not exist.  The problem
highlighted by this is not limited to spring-cloud; any library and
any class can exhibit this behavior.  This change adds a guard around
the cloud services test to ensure that a class can even be loaded
before we go looking for beans of that type.  In the case that the
class cannot be loaded, it is assumed that it is not being used.

[#72625668]