Play auto-reconfiguration missing conf files
Previously when auto-reconfiguration worked against a Playy
application it would correctly detect existing configuration files and
process them.  However, if a configuration file was listed that did
not exist, it would throw an exception.  This behavior did not match
the behavior of Play itself.  This change ensures that a missing
configuration file is ignored rather than causing an exception.

[#78905842][resolves #41]