Add CloudProfileApplicationListener and use that for Spring Boot
Instead of using the CloudProfileApplicationContextInitializer, which
is fine in a container application but comes too late in the lifecycle
of a Spring Boot app, we register an ApplicationListener to activate
the "cloud" profile.

The Ordered implementation ensures that this listener is processed
before Spring Cloud needs to know about profiles as well.