Auto-Reconfiguration 2.x
This change is a rewrite of the auto-reconfiguration support in the Java
Buildpack.  Previously there was a quite a lot of code in the auto-
reconfiguration project that examined the internals of bean definitions and
the various ways those bean definitions might be constructed to replace
detailed portions.  The new view in 2.x is that we're going to do auto-
reconfiguration at the bean level and no lower.  This is easily the most
typical way that auto-reconfiguration was used and has the added advantage of
being attuned to how Spring Boot creates beans.

In addition to massively simplifying the code base by simply matching a bean
type and a service type and replacing one for the other, this change also
allows us to easily expand the auto-reconfiguration to the complete set of
default service types understood by Spring Cloud Connectors.  To complete the
list we've added Cassandra and SMTP support.

We've also massively simplified the test matrix for the project by adopting
Spring.io platform releases (i.e. platform-bom) as the set of compatible
versions that we will test against.  The CI system, and our planned support
starts at Athens-RELEASE (Spring Boot 1.4) and will cover every subsequent
release (e.g. Brussels-RELEASE, Cairo-RELEASE), as well as every service
release (e.g. Athens-SR6, Brussels-SR5).  This starts by chopping our test
matrix to a third of its original size (but possibly out growing it over the
long term) but giving us a much more robust combination of dependencies (since
they're the dependencies most likely to be used by customers) to test against.
It is not only possible, but likely, that auto-reconfiguration will work
further back than that (thank you Spring Data for being so compatible), but we
will not be testing against it.

Finally, this change includes the source code into a shaded source code JAR.
Note that some of the pre-shaded code from Spring Cloud Connectors is not
included in its source jar and therefore is missing.

[resolves #48][resolves #49][resolves #51][resolves #52]