Move Using Cloud Services Check
Previously, the check to determine if the code was running with cloud services
already defined was performed during the ApplicationContextInitialization.
This was way too early to do this check (some bean definitions needed to be in
place).  This change moves that check to the BeanFactoryPostProcessor step as
it was previously.